bash completion: fixed special character handling
I experienced problems using the bash-completion for mpc. When trying
to autocomplete a partial pathname, i got errors from bash because
there was one song in the autocompletion list with a "'" in it.  The
autocompletion worked, but there was lots of ugly output printed after
pressing <tab>.  I fixed this problem, also the list of possible
completions is now escaped right after running `mpc tab`.  The
escaping was done after running `compgen` before, and that is too late
because compgen itself is a bash-builtin, and so itself spits out
reserved characters.

I added a patch to this report, but am not shure wether this is the
right place to post this report to, so if not maybe someone can tell
me who's working at bash-completion for mpc, I'm using Gentoo-Linux.

[mk: this patch was posted anonymously to the MPD bug tracker, see
http://musicpd.org/mantis/view.php?id=510]