command: support output names in enable/disable
This commit changes enable and disable commands so that they accept
not only output number but also its name.  This way, one does not
need to remember the number of the output but instead use its name.
It also makes user command work even if number and order of outputs
in configuration file changes (as long as names stay the same).