New command seekthrough -- seek through playlist
This enables to seek forward or backward by an arbitrary duration
relative to the current position in the track. If the duration stays
within the limit of the current track, the result is the same as command
seek. If the duration exceeds the limits of the current track (forward
or backward), it continues seeking into next/previous songs until
accumulating the exact duration asked, or reaching the end/beginning of
the playlist.

Durations in percentages or absolute position, which exist for command
seek, do not make sense for this command.

If music was initially paused, the paused status is restored in the end.

An arbitrary limit of 100 iterations is set to avoid an infinite loops
(could happen if duration of the song is very short, rounding to zero,
and player is set to repeat).