Avoid underlines in generated manpage
This change modifies certain inline markup to fence it as
\`\`code\`\`, instead of using the \`default role\`. When
generated as a man page, text in the default role is
underlined, which can cause readability issues on the
terminal if the fenced text includes underscores or
characters like square brackets,both of which overlap with
the underline. As \`\`code\`\` the fenced text will be
boldfaced instead of underlined.