doc: make upload target dependency on rsync optional
Currently, rsync is an unconditional dependency and checked during
`meson configure`. As such, the build will fail if rsync is not
installed which is probably not what was intended here.

From the meson documentation:

	Meson will automatically insert the appropriate dependencies on
	targets and files listed in this keyword [the command] argument.

This commit fixes the unconditional dependency on rsync with an explicit
find_program invocation with `required: false`. Also wrap the
custom_target in an if statement since it is not allowed to use
non-found external programs in `command`.