Fix mpc status ''
Currently when calling mpc status with a empty string argument, the
format_object2 returns a NULL string, which leads to a segmentation fault when
the status gets printed. The implemented fix adds a check before printing the
returned status from format_song. Other callings of format_object2 seem to not
be affected by this behaviour.

Signed-off-by: shirenn <shirenn@crans.org>